Requests::Version: update version number
Oops.. looks like we forgot something for the release. Only just noticed it when I checked the file diff for the WP Core patch.

As the version constant is generally used to verify the version for using certain features of a package, does this warrant a `1.8.1` release ?

And if so, should I update the version number to `1.8.1` in this PR ?